pyproject.toml
[project]
name = "pulse"
version = "0.1.0"
description = "Agentic AI System"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "tenacity>=8.2.0",
    "python-multipart>=0.0.6",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "langchain-mcp-adapters>=0.2.2",
    "langchain>=1.2.13",
    "langchain-openai>=1.1.12",
    "langchain-google-vertexai>=3.2.2",
    "langchain-google-genai>=4.2.1",
    "langchain-community>=0.4.1",
    "google-genai>=0.8.0",
    "feedparser>=6.0.12",
    "arxiv>=2.4.1",
    "langgraph>=0.2.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "trafilatura>=2.0.0",
    "ddgs>=9.13.1",
    "resend>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
pulse = "main:main"
